Add the similarity package to the base library
The relative similarity of entity references  is one of two well
established ways to tackle entity resolution. A key data type in
assessing relative similarity is the similarity graph. This graph
is a directed graph constructed using a matcher and the thresholds
specified by the Fellegi-Sunter model of entity resolution: the
maximum non-match threshold (below which everything is a non-match)
and the minimum match threshold (above which everything is a
match).

The similarity package provides:

- a Matcher protocol that can be adhered to by any matcher
implementation
- the SimilarityGraph data type which implements the similarity
graph using any object that implements the Matcher protocol.